Fixing endpoints to contain year parameter

+++ server/controllers/billController.js
@@ -20,19 +20,29 @@ const checkCategoryExists = async (username, category) => {
 const getBills = async (req, res) => {
     // authenticate signed in user
 
+    const { year } = req.params;
+    if (!year || !(year in yrToModel)) {
+        return res.status(404).json({error: 'Year not provided or year does not exist'})
+    }
+
     // get all bills for owner
-    const bills = await yrToModel['2023'].find({owner: 'Tyler'}).sort({ date: -1 });
+    const bills = await yrToModel[year].find({owner: 'Tyler'}).sort({ date: -1 });
     res.status(200).json(bills);
 }
 
 // Get all bills documents for specific category
-const getBillsCategory = async (req, res) => {
+const getBillsByCategory = async (req, res) => {
     // authenticate signed in user
 
-    const { category } = req.params;
-
+    const { category, year } = req.params;
+    
+    
     if (!checkCategoryExists) {
-        return res.status(404).json({error: "Category does not exist"})
+        return res.status(404).json({error: 'Category does not exist'})
+    }
+    
+    if (!year || !(year in yrToModel)) {
+        return res.status(404).json({error: 'Year not provided or year does not exist'})
     }
 
     // get all bills for owner
@@ -42,18 +52,22 @@ const getBillsCategory = async (req, res) => {
 
 // Get single bill document in collection
 const getBill = async (req, res) => {
-    const { id } = req.params;
+    const { id, year } = req.params;
 
     // validate id
     if (!isValid(id)) {
-        return res.status(404).json({error: "No such bill"});
+        return res.status(404).json({error: 'Bill does not exist'});
+    }
+    
+    if (!year || !(year in yrToModel)) {
+        return res.status(404).json({error: 'Year not provided or year does not exist'})
     }
 
     // authenticate signed in user
 
     const bills = await yrToModel['2023'].findById(id);
     if (!bills) {
-        return res.status(404).json({error: "No such bill"});
+        return res.status(404).json({error: 'No such bill'});
     }
     res.status(200).json(bills);
 }
@@ -63,7 +77,7 @@ const createBill = async (req, res) => {
     const { owner, category, subcategory, date, amount, description } = req.body;
 
     if (!checkCategoryExists) {
-        return res.status(404).json({error: "Category does not exist"})
+        return res.status(404).json({error: 'Category does not exist'})
     }
 
     try {
@@ -75,28 +89,36 @@ const createBill = async (req, res) => {
 }
 
 const deleteBill = async (req, res) => {
-    const { id } = req.params;
+    const { id, year } = req.params;
 
     // validate id
     if (!isValid(id)) {
-        return res.status(404).json({error: "No such bill"});
+        return res.status(404).json({error: 'Bill does not exist'});
+    }
+
+    if (!year || !(year in yrToModel)) {
+        return res.status(404).json({error: 'Year not provided or year does not exist'})
     }
 
     // authenticate signed in user
 
     const bills = await yrToModel['2023'].findOneAndDelete({_id: id});
     if (!bills) {
-        return res.status(404).json({error: "No such bill"});
+        return res.status(404).json({error: 'No such bill'});
     }
     res.status(200).json(bills);
 }
 
 const updateBill = async (req, res) => {
-    const { id } = req.params;
+    const { id, year } = req.params;
 
     // validate id
     if (!isValid(id)) {
-        return res.status(404).json({error: "No such bill"});
+        return res.status(404).json({error: 'Bill does not exist'});
+    }
+
+    if (!year || !(year in yrToModel)) {
+        return res.status(404).json({error: 'Year not provided or year does not exist'})
     }
 
     // authenticate signed in user
@@ -106,38 +128,43 @@ const updateBill = async (req, res) => {
 
     const bills = await yrToModel['2023'].findOneAndUpdate({_id: id}, {...req.body});
     if (!bills) {
-        return res.status(404).json({error: "No such bill"})
+        return res.status(404).json({error: 'No such bill'})
     }
     res.status(200).json(bills);
 }
 
 const resetCollection = async (req, res) => {
-    const { pass, collection } = req.body;
+    const { year } = req.params;
 
-    if (!pass || pass !== adminPass) {
-        return res.status(401).json({error: "Unauthorized access"});
+    if (!year || !(year in yrToModel)) {
+        return res.status(404).json({error: 'Year not provided or year does not exist'})
     }
 
-    if (!(collection in yrToModel)) {
-        return res.status(404).json({error: "Collection not provided or collection does not exist"});
+    const { pass } = req.body;
+
+    if (!pass || pass !== adminPass) {
+        return res.status(401).json({error: 'Unauthorized access'});
+    }
+
+    if (!year || !(year in yrToModel)) {
+        return res.status(404).json({error: 'Collection not provided or collection does not exist'});
     } 
 
     try {
-        await yrToModel[collection].deleteMany({});
-        res.status(200).json({message: `${collection} successfully reset`});
+        await yrToModel[year].deleteMany({});
+        res.status(200).json({message: `${year} successfully reset`});
     }
     catch (error) {
         // Log error for debugging purposes
-        console.error("Error resetting collection:", error);
-
-        return res.status(500).json({error: "Resetting collection failed"});
+        console.error('Error resetting collection:', error);
+        return res.status(500).json({error: 'Resetting collection failed'});
     }
 }
 
 // Export the functions
 module.exports = {
     getBills,
-    getBillsCategory,
+    getBillsByCategory,
     getBill,
     createBill,
     deleteBill,
